% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiKink-package.R
\docType{package}
\name{MultiKink-package}
\alias{MultiKink}
\alias{MultiKink-package}
\title{MultiKink: Estimation and Inference for Multi-Kink Quantile Regression}
\description{
Estimation and inference for multiple kink quantile regression for longitudinal data and the i.i.d data. A bootstrap restarting iterative segmented quantile algorithm is proposed to estimate the multiple kink quantile regression model conditional on a given number of change points. The number of kinks is also allowed to be unknown. In such case, the backward elimination algorithm and the bootstrap restarting iterative segmented quantile algorithm are combined to select the number of change points based on a quantile BIC. For longitudinal data, we also develop the GEE estimator to incorporate the within-subject correlations. A score-type based test statistic is also developed for testing the existence of kink effect. The package is based on the paper, ``Wei Zhong, Chuang Wan and Wenyang Zhang (2022). Estimation and inference for multikink quantile regression, JBES'' and ``Chuang Wan, Wei Zhong, Wenyang Zhang and Changliang Zou (2022). Multi-kink quantile regression for longitudinal data with application to progesterone data analysis, Biometrics".
}
\author{
\strong{Maintainer}: Chuang Wan \email{wanchuanghnu@126.com}

Authors:
\itemize{
  \item Wei Zhong \email{wzhong@xmu.edu.cn}
}

}
\keyword{internal}
