% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triceps}
\alias{triceps}
\title{Triceps Skinfold Thickness Dataset}
\format{
A data frame with 892 observations on the following 3 variables:
\describe{
  \item{\code{age}}{Age of respondents.}
  \item{\code{lntriceps}}{Log of the triceps skinfold thickness.}
  \item{\code{triceps}}{Triceps skinfold thickness.}
}
}
\source{
Cole T.J., Green P.J. (1992). Smoothing reference centile curves: the LMS method and penalized likelihood. Statistics in medicine, 11(10): 1305-1319.
}
\usage{
data(triceps)
}
\description{
The data are derived from an anthropometric study of 892 females under 50 years in three Gambian villages in West Africa.
}
\examples{
data(triceps)
## maybe str(triceps) ...
}
\references{
Cole T.J., Green P.J. (1992). Smoothing reference centile curves: the LMS method and penalized likelihood. Statistics in medicine, 11(10): 1305-1319.
}
\keyword{datasets}
